test: add common module
