+++ tests/common/mod.rs
@@ -1,0 +1,16 @@
+use ril::prelude::*;
+
+pub const COLORS: [Rgb; 12] = [
+    Rgb::new(255, 0, 0),
+    Rgb::new(255, 128, 0),
+    Rgb::new(255, 255, 0),
+    Rgb::new(128, 255, 0),
+    Rgb::new(0, 255, 0),
+    Rgb::new(0, 255, 128),
+    Rgb::new(0, 255, 255),
+    Rgb::new(0, 128, 255),
+    Rgb::new(0, 0, 255),
+    Rgb::new(128, 0, 255),
+    Rgb::new(255, 0, 255),
+    Rgb::new(255, 0, 128),
+];
