+++ tests/test_png.rs
@@ -1,20 +1,7 @@
+mod common;
+use common::COLORS;
 use ril::prelude::*;
 use std::time::Duration;
-
-pub const COLORS: [Rgb; 12] = [
-    Rgb::new(255, 0, 0),
-    Rgb::new(255, 128, 0),
-    Rgb::new(255, 255, 0),
-    Rgb::new(128, 255, 0),
-    Rgb::new(0, 255, 0),
-    Rgb::new(0, 255, 128),
-    Rgb::new(0, 255, 255),
-    Rgb::new(0, 128, 255),
-    Rgb::new(0, 0, 255),
-    Rgb::new(128, 0, 255),
-    Rgb::new(255, 0, 255),
-    Rgb::new(255, 0, 128),
-];
 
 #[test]
 fn test_static_png() -> ril::Result<()> {
