+++ tests/test_webp.rs
@@ -1,9 +1,9 @@
 #![cfg(feature = "webp")]
-mod test_png;
 
+mod common;
+use common::COLORS;
 use ril::prelude::*;
 use std::time::Duration;
-use test_png::COLORS;
 
 #[test]
 fn test_static_webp_encode() -> ril::Result<()> {
